fix(EntityModel): move configurations to EntityModel project
FinanceDbContext needs to be able to access these configurations, so they
cannot reside in the EntityMigration.
