+++ NjordinSight.EntityModel/Context/Configurations/AccountObjectConfiguration.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NjordinSight.EntityMigration.Configurations
+namespace NjordinSight.EntityModel.Context.Configurations
 {
     public class AccountObjectConfiguration : IEntityTypeConfiguration<AccountObject>
     {
